Split NextJS static image logic into its own monorepo package
