+++ editor/package.json
@@ -17,20 +17,21 @@
   },
   "dependencies": {
     "@sindresorhus/slugify": "^2.2.1",
-    "@tanstack/react-query": "^5.25.0",
+    "@tanstack/react-query": "^5.28.4",
     "bcrypt": "^5.1.1",
     "clsx": "^2.1.0",
     "component-library": "^1.0.0",
-    "date-fns": "^3.3.1",
+    "date-fns": "^3.6.0",
     "entities": "^4.5.0",
     "flexsearch": "^0.7.43",
     "format-quantity": "^3.0.0",
     "fraction.js": "^4.3.7",
-    "lmdb": "^2.9.4",
+    "lmdb": "^3.0.0",
     "lodash": "^4.17.21",
-    "markdown-to-jsx": "^7.4.1",
+    "markdown-to-jsx": "^7.4.3",
     "next": "14.1.3",
     "next-auth": "^5.0.0-beta.9",
+    "next-static-image": "^0.0.1",
     "react": "^18",
     "react-dom": "^18",
     "recipes-collection": "^1.0.0",
@@ -40,17 +41,17 @@
   "devDependencies": {
     "@testing-library/cypress": "^10.0.1",
     "@types/bcrypt": "^5.0.2",
-    "@types/lodash": "^4.14.202",
+    "@types/lodash": "^4.17.0",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "autoprefixer": "^10.4.18",
-    "cypress": "^13.6.6",
-    "eslint": "^8.56.0",
+    "cypress": "^13.7.0",
+    "eslint": "^8.57.0",
     "eslint-config-next": "14.1.3",
     "eslint-plugin-cypress": "^2.15.1",
     "eslint-plugin-mocha": "^10.4.1",
-    "eslint-plugin-react": "^7.34.0",
+    "eslint-plugin-react": "^7.34.1",
     "postcss": "^8",
     "prettier": "3.2.5",
     "start-server-and-test": "^2.0.3",
