+++ next-static-image/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "component-library",
-  "version": "1.0.0",
+  "name": "next-static-image",
+  "version": "0.0.1",
   "description": "",
   "main": "index.js",
   "scripts": {
@@ -11,7 +11,10 @@
   "license": "ISC",
   "dependencies": {
     "clsx": "^2.1.0",
-    "next": "14.1.3",
-    "markdown-to-jsx": "^7.4.3"
+    "markdown-to-jsx": "^7.4.3",
+    "next": "14.1.3"
+  },
+  "devDependencies": {
+    "@types/node": "^20"
   }
 }
