+++ recipes-collection/package.json
@@ -16,6 +16,7 @@
     "fraction.js": "^4.3.7",
     "lodash": "^4.17.21",
     "next": "14.1.3",
+    "next-static-image": "^0.0.1",
     "plaiceholder": "^3.0.0",
     "react": "^18",
     "sharp": "0.33.2",
