+++ website/package.json
@@ -17,17 +17,17 @@
   },
   "dependencies": {
     "@sindresorhus/slugify": "^2.2.1",
-    "@tanstack/react-query": "^5.25.0",
+    "@tanstack/react-query": "^5.28.4",
     "clsx": "^2.1.0",
     "component-library": "^1.0.0",
-    "date-fns": "^3.3.1",
+    "date-fns": "^3.6.0",
     "flexsearch": "^0.7.43",
     "format-quantity": "^3.0.0",
     "fraction.js": "^4.3.7",
-    "lmdb": "^2.9.4",
+    "lmdb": "^3.0.0",
     "lodash": "^4.17.21",
-    "markdown-to-jsx": "^7.4.1",
-    "netlify-cli": "^17.19.0",
+    "markdown-to-jsx": "^7.4.3",
+    "netlify-cli": "^17.19.5",
     "next": "14.1.3",
     "react": "^18",
     "react-dom": "^18",
@@ -38,15 +38,15 @@
   "devDependencies": {
     "@testing-library/cypress": "^10.0.1",
     "@types/bcrypt": "^5.0.2",
-    "@types/lodash": "^4.14.202",
+    "@types/lodash": "^4.17.0",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
     "autoprefixer": "^10.4.18",
-    "cypress": "^13.6.6",
+    "cypress": "^13.7.0",
     "eslint": "^8",
     "eslint-config-next": "14.1.3",
-    "eslint-plugin-react": "^7.34.0",
+    "eslint-plugin-react": "^7.34.1",
     "postcss": "^8",
     "serve": "^14.2.1",
     "start-server-and-test": "^2.0.3",
